+++ templates/favorites/item.hbs
@@ -1,38 +1,76 @@
-<li class="item flexrow fav-item" data-item-id="{{item._id}}">
-    <div class="item-image flexrow rollable" style="background-image: url({{item.img}})"></div>
-    <div class="item-shortcuts flexrow">
-        {{#if item.hasAttack}} <span class="tag rollable attack" title="Roll Attack"><i class="ra ra-sword"></i></span> {{/if}}
-        {{#if item.hasDamage}} <span class="tag rollable damage" title="Roll Damage"><i class="fas fa-tint"></i></span> {{/if}}
-    </div>
-    <div class="item-name flexrow">        
-        <h4>{{item.name}}{{item.spellPrepMode}}</h4>
-    </div>
-    <div class="labels">
-        {{#each item.favLabels as |label id|}}
-            <span class="tag">{{label}}</span>
-        {{/each}}
-    </div>
-    {{#if item.data.quantity}} 
-        <div class="item-detail count">
-            <input class="item-count" data-path="data.quantity" type="text" value="{{item.data.quantity}}" {{#if item.editable}}{{else}}disabled{{/if}}> 
-        </div>
+<li class="item flexrow fav-item{{#if item.data.attuned}} attuned{{/if}}{{#if item.data.equipped}} equipped{{/if}}{{#if item.data.preparation.prepared}} prepared {{/if}}" data-item-id="{{item._id}}">
+  <div class="item-image flexrow rollable" style="background-image: url({{item.img}})"></div>
+  <div class="item-name flexrow">        
+    <h4 title="Toggle Info">
+      <i class="fas fa-info-circle"></i>
+      {{item.name}}
+      <!-- {{item.spellPrepMode}} -->
+    </h4>
+    {{#if item.data.quantity}}
+      <span class="item-quantity{{#if item.isStack}} isStack{{/if}}">
+        (<input class="item-count" data-path="data.quantity" type="text" value="{{item.data.quantity}}" maxlength="3" {{#if item.editable}}{{else}}disabled{{/if}}>)
+      </span>
     {{/if}}
-    <div class="item-detail uses">
-        {{#if item.data.uses.max}}
-            <span style="display:inline;">(</span>
-                <input data-path="data.uses.value" type="text" value="{{item.data.uses.value}}" {{#if item.editable}}{{else}}disabled{{/if}}>
-            <span style="display:inline;">/</span>
-                <input data-path="data.uses.max" type="text" value="{{item.data.uses.max}}" {{#if item.editable}}{{else}}disabled{{/if}}>        
-            <span style="display:inline;">)</span>
-        {{else}}
-            {{#if item.editable}}<a class="addCharges" value="Add Charges">Add Charges</a>{{/if}}
-        {{/if}}
+  </div>
+
+  {{#if item.isItem}}
+  <div class="state-display item-equipped">
+    <a class="item-control item-toggle {{item.toggleClass}}" data-path="data.equipped" title="{{item.toggleTitle}}">
+      <i class="fas fa-user-alt-slash"></i>
+      <i class="fas fa-user-alt"></i>
+    </a>
+  </div>
+  {{/if}}
+
+  {{#if item.canPrep}}
+  <div class="state-display spell-details spell-prepared">
+    <a class="item-control item-toggle {{item.toggleClass}}" data-path="data.preparation.prepared" title="{{item.toggleTitle}}">
+      <i class="fas fa-book"></i>
+    </a>
+  </div>
+  {{/if}}
+
+  {{#if item.isItem}}
+  <div class="item-detail item-uses">
+    {{#if item.data.uses.max}}
+    <input data-path="data.uses.value" type="text" value="{{item.data.uses.value}}" {{#if item.editable}}{{else}}disabled{{/if}}>
+    / <input  class="uses-max" data-path="data.uses.max" type="text" value="{{item.data.uses.max}}" {{#if item.editable}}{{else}}disabled{{/if}}>
+    {{else}}
+    {{#if item.editable}}<a class="addCharges" value="Add">Add</a>{{/if}}
+    {{/if}}
+  </div>
+  {{/if}}
+  {{#if item.isFeat}}
+  <div class="item-detail item-uses">
+    {{#if item.data.uses.max}}
+    <input data-path="data.uses.value" type="text" value="{{item.data.uses.value}}" {{#if item.editable}}{{else}}disabled{{/if}}>
+    / <input class="uses-max" data-path="data.uses.max" type="text" value="{{item.data.uses.max}}" {{#if item.editable}}{{else}}disabled{{/if}}>
+    {{else}}
+    {{#if item.editable}}<a class="addCharges" value="Add">Add</a>{{/if}}
+    {{/if}}
+  </div>
+  {{/if}}
+
+  {{#if item.spellComps}}
+    <div class="spell-detail spell-comps">
+      {{item.spellComps}}
+      {{#if item.spellCon}}<span class="spell-component C">C</span>{{/if}}
+      {{#if item.spellRit}}<span class="spell-component R">R</span>{{/if}}
     </div>
-    <div class="item-detail item-controlls">
-        {{#if item.canPrep}}<a class="item-control item-toggle {{item.toggleClass}}" data-path="data.equipped" title="{{item.toggleTitle}}"><i class="fas fa-sun"></i></a>{{/if}}
-        {{#if item.isItem}}<a class="item-control item-toggle {{item.toggleClass}}" data-path="data.preparation.prepared" title="{{item.toggleTitle}}"><i class="fas fa-shield-alt"></i></a>{{/if}}
-        
-        <a class="item-control item-edit" title="Edit Item" > <i class="fas fa-edit" ></i></a>
-        <a class="item-control item-fav" title="Remove from favourites" > <i class="fas fa-sign-out-alt" ></i></a>
-    </div>
+  {{/if}}
+
+  <div class="item-detail item-action">
+    {{#each item.favLabels as |label id|}}
+      <span>{{label}}</span>
+    {{/each}}
+  </div>
+
+  <div class="item-detail item-controls">
+    <a class="item-control item-edit" title="Edit Item">
+      <i class="fas fa-pencil-alt"></i>
+    </a>
+    <a class="item-control item-fav" title="Remove from Favourites">
+      <i class="fas fa-bookmark"></i>
+    </a>
+  </div>
 </li>
