+++ templates/favorites/template.hbs
@@ -1,48 +1,94 @@
-<div class="tab inventory-list favourite" data-group="primary" data-tab="favourite">
-    {{#if favItems}}
-    <ol class="fav-item">
-        <li class="item inventory-header flexrow">
-            <h3>Item</h3>
-            <div class="item-detail count">Count</div>
-            <div class="item-detail uses">Uses</div>
-            <div class="item-detail item-controlls">Control</div>
-        </li>
-        {{#each favItems as |item id|}}
-            {{> "modules/favtab/templates/item.hbs" item=item}}
-        {{/each}}
+<div class="favorites">
+  {{#if favItems}}
+    <ol class="inventory-list fav-item">
+      <li class="item inventory-header flexrow">
+        <h3>Items <span title="Quantity">(qty)</span></h3>
+        <div class="item-detail item-uses"><i class="fas fa-bolt"></i></div>
+        <div class="item-detail item-action"><i class="fas fa-hourglass"></i></div>
+      </li>
+      {{#each favItems as |item id|}}
+       {{> "modules/tidy5e-sheet/templates/favorites/item.hbs" item=item}}
+      {{/each}}
     </ol>
-    {{/if}}
-    {{#if favFeats}}
-    <ol class="fav-feat">
-        <li class="item inventory-header"><h3>Feats</h3></li>
-        {{#each favFeats as |feat lvl|}}
-            {{> "modules/favtab/templates/item.hbs" item=feat}}
-        {{/each}}
+  {{/if}}
+  {{#if favFeats}}
+    <ol class="inventory-list fav-feat">
+      <li class="item inventory-header flexrow">
+        <h3>Features</h3>
+        <div class="item-detail item-uses"><i class="fas fa-bolt"></i></div>
+        <div class="item-detail item-action"><i class="fas fa-hourglass"></i>
+      </li>
+      {{#each favFeats as |feat lvl|}}
+      {{> "modules/tidy5e-sheet/templates/favorites/item.hbs" item=feat}}
+      {{/each}}
     </ol>
-    {{/if}}
-    {{#if favSpells}}
-    <ol class="fav-spell">
-        {{#each favSpells as |section lvl|}}
-            {{#if section.spells}}
-                {{#if lvl}}
-                    <li class="item inventory-header flexrow">                        
-                        <h3>{{#if section.isCantrip}}Cantrip{{else}}Spell Level {{lvl}}{{/if}}</h3>
-                        {{#if section.isCantrip}}{{else}}
-                        <span class="spell-slots">
-                            <input type="text" ${inputStyle} data-target="data.spells.spell{{lvl}}.value" value="{{section.value}}" placeholder="0"/>
-                            /
-                            <input type="text" ${inputStyle} data-target="data.spells.spell{{lvl}}.max" value="{{section.max}}" placeholder="0"/>
-                        </span>
-                        {{/if}}
-                    </li>
+  {{/if}}
+
+  {{#if favSpellsPrepMode}}
+    {{#each favSpellsPrepMode as |section lvl|}}
+      <ol class="inventory-list fav-spell">
+        {{#if section.spells}}
+          {{#if lvl}}
+            <li class="item inventory-header flexrow">                        
+              <h3>
+              {{#if section.isAlways}}Always Prepared
+              {{else if section.isAtWill}}At Will
+              {{else if section.isInnate}}Innate Spellcasting
+              {{else if section.isPact}}Pact Magic
+              {{/if}}
+              </h3>
+              {{#if section.isPact}}
+                <span class="spell-slots">
+                  <input type="text" ${inputStyle} data-target="data.spells.spell{{lvl}}.value" value="{{section.value}}" placeholder="0"/>
+                  /
+                  <input class="spell-max" type="text" ${inputStyle} data-target="data.spells.spell{{lvl}}.max" value="{{section.max}}" placeholder="0"/>
+                </span>
                 {{else}}
-                    <li class="item inventory-header"><h3>Cantrips</h3></li>
-                {{/if}}
-            {{#each section.spells as |spell id|}}
-                {{> "modules/favtab/templates/item.hbs" item=spell}}
-            {{/each}}
-            {{/if}}
-        {{/each}}
-    </ol>
-    {{/if}}
+                <!-- <span class="spell-slots">
+                  &ndash; / &ndash;
+                </span> -->
+              {{/if}}
+              <div class="item-detail item-action"><i class="fas fa-hourglass"></i>
+            </li>
+          {{/if}}
+          {{#each section.spells as |spell id|}}
+            {{> "modules/tidy5e-sheet/templates/favorites/item.hbs" item=spell}}
+          {{/each}}
+        {{/if}}
+      </ol>
+    {{/each}}
+  {{/if}}
+  {{#if favSpells}}
+    {{#each favSpells as |section lvl|}}
+      <ol class="inventory-list fav-spell">
+        {{#if section.spells}}
+          {{#if lvl}}
+            <li class="item inventory-header flexrow">                        
+              <h3>
+              {{#if section.isCantrip}}Cantrips
+              {{else if section.isAlways}}Always Prepared
+              {{else if section.isAtWill}}At Will
+              {{else if section.isInnate}}Innate Spellcasting
+              {{else if section.isPact}}Pact Magic
+              {{else}}Spell Level {{lvl}}{{/if}}
+              </h3>
+              {{#if section.isCantrip}}
+                <!-- <span class="spell-slots">∞/∞</span> -->
+                {{else}}
+                <span class="spell-slots">
+                  <input type="text" ${inputStyle} data-target="data.spells.spell{{lvl}}.value" value="{{section.value}}" placeholder="0"/>
+                  /
+                  <input class="spell-max" type="text" ${inputStyle} data-target="data.spells.spell{{lvl}}.max" value="{{section.max}}" placeholder="0"/>
+                </span>
+              {{/if}}
+              <div class="item-detail item-action"><i class="fas fa-hourglass"></i>
+            </li>
+          {{/if}}
+          {{#each section.spells as |spell id|}}
+            {{> "modules/tidy5e-sheet/templates/favorites/item.hbs" item=spell}}
+          {{/each}}
+        {{/if}}
+      </ol>
+    {{/each}}
+  {{/if}}
 </div>
