--- templates/item.hbs
@@ -1,23 +1,0 @@
-<li class="item flexrow fav-item{{#if item.data.attuned}} attuned{{/if}}{{#if item.data.equipped}} equipped{{/if}}" data-item-id="{{item._id}}">
-	<div class="item-name flexrow rollable">
-		<div class="item-image" style="background-image: url({{item.img}})"></div>
-		<h4 title="Toggle Info">{{item.name}}{{#if item.showquant}} ({{item.data.quantity}}){{/if}}</h4>
-	</div>
-
-  {{#if item.spellComps}}
-		<div class="item-detail uses">
-			{{item.spellComps}}
-			{{#if item.spellCon}}<i class='fas fa-copyright'></i>{{/if}}
-			{{#if item.spellRit}}<i class='fas fa-registered'></i>{{/if}}
-		</div>
-	{{/if}}
-	<div class="item-detail item-action">
-    {{#if item.data.activation.type }}
-    {{item.action}}
-    {{/if}}
-  </div>	
-	<div class="item-detail item-controls">
-		<a class="item-control item-edit" title="Edit Item"><i class="fas fa-pencil-alt"></i></a>
-		<a class="item-control item-fav" title="Remove from Favorites"><i class="fas fa-bookmark"></i></a>
-	</div>
-</li>
