+++ templates/parts/tidy5e-spellbook.html
@@ -31,9 +31,9 @@
         </a>
         {{/if}}
         {{ else }}
-        <span class="spell-slots">{{{section.uses}}}</span>
+        <!-- <span class="spell-slots">{{{section.uses}}}</span>
         <span class="sep"> / </span>
-        <span class="spell-max">{{{section.slots}}}</span>
+        <span class="spell-max">{{{section.slots}}}</span> -->
         {{/if}}
       </div>
