+++ templates/tidy5e-sheet.html
@@ -43,7 +43,7 @@
       			<p class="lvl-6">Death</p>
       		</div>
       		<div class="exhaustion">
-      			<div class="exhaustion-icon"><i class="far fa-tired"></i></div>
+      			<div class="exhaustion-icon"><i class="far fa-grin"></i><i class="far fa-smile"></i><i class="far fa-meh"></i><i class="far fa-frown"></i><i class="far fa-frown-open"></i><i class="far fa-tired"></i><i class="far fa-dizzy"></i></div>
       			<ul class="exhaust-level">
             	<li data-elvl="0">0</li>
             	<li data-elvl="1">1</li>
@@ -261,7 +261,7 @@
 							<h4>Character Traits</h4>
 							<div class="toggle-icon">
 								<i class="fas fa-caret-down"></i>
-								<i class="fas fa-caret-up"></i>
+								<i class="fas fa-caret-right"></i>
 							</div>
 						</header>
 						{{> "modules/tidy5e-sheet/templates/parts/tidy5e-traits.html"}}
@@ -273,7 +273,7 @@
 							<h4>Favorites</h4>
 							<div class="toggle-icon">
 								 <i class="fas fa-caret-down"></i>
-								 <i class="fas fa-caret-up"></i>
+								 <i class="fas fa-caret-right"></i>
 							</div>
 						</header>
 						<div class="favorites-target"></div>
