+++ tidy5e-sheet.js
@@ -3,8 +3,12 @@ import ActorSheet5e from "../../systems/dnd5e/module/actor/sheets/base.js";
 import ActorSheet5eCharacter from "../../systems/dnd5e/module/actor/sheets/character.js";
 
 import { preloadTidy5eHandlebarsTemplates } from "./templates/tidy5e-templates.js";
+import { addFavorites } from "./tidy5e-favorites.js";
+
+let scrollPos = 0;
 
 export class Tidy5eSheet extends ActorSheet5eCharacter {
+	
 	get template() {
 		if ( !game.user.isGM && this.actor.limited ) return "modules/tidy5e-sheet/templates/tidy5e-sheet-ltd.html";
 		return "modules/tidy5e-sheet/templates/tidy5e-sheet.html";
@@ -25,6 +29,18 @@ export class Tidy5eSheet extends ActorSheet5eCharacter {
 
 	activateListeners(html) {
 		super.activateListeners(html);
+
+		// store Scroll Pos
+		let attributesTab = html.find('.tab.attributes');
+		attributesTab.scroll(function(){
+			scrollPos = $(this).scrollTop();
+		});
+		let tabNav = html.find('a.item:not([data-tab="attributes"])');
+		tabNav.click(function(){
+			scrollPos = 0;
+			attributesTab.scrollTop(scrollPos);
+		});
+
 		// toggle item delete protection
 		html.find('.tidy5e-delete-toggle').click(async (event) => {
 			event.preventDefault();
@@ -73,6 +89,22 @@ export class Tidy5eSheet extends ActorSheet5eCharacter {
 
 	}
 }
+
+// store Scroll Position
+// function storeScrollPos(html){
+// 	let attributesTab = html.find('.tab.attributes');
+// 	if(attributesTab.hasClass('active')){
+// 		attributesTab.scroll(function(){
+// 			scrollPos = $(this).scrollTop();
+// 		});
+// 	} 
+// 	let tabs = html.find('.tab:not(.attributes)');
+// 	tabs.click(function(){
+// 	console.log($(this));
+// 		scrollPos = 0;
+// 		attributesTab.scrollTop(scrollPos);
+// 	});
+// }
 
 // Migrate Traits to default dnd5e data
 async function migrateTraits(app, html, data) {
@@ -171,178 +203,6 @@ async function hidePortraitButtons(app, html, data){
 	};
 }
 
-// The following function is adapted for the Tidy5eSheet from the Favorites Item
-// Tab Module created for Foundry VTT - by Felix Müller (Felix#6196 on Discord).
-// It is licensed under a Creative Commons Attribution 4.0 International License
-// and can be found at https://github.com/syl3r86/favtab.
-
-async function addFavorites(app, html, data) {
-	let favItems = [];
-	let favFeats = [];
-	let favSpells = {
-		0: { isCantrip: true, spells: [] },
-		1: { spells: [], value: data.actor.data.spells.spell1.value, max: data.actor.data.spells.spell1.max },
-		2: { spells: [], value: data.actor.data.spells.spell2.value, max: data.actor.data.spells.spell2.max },
-		3: { spells: [], value: data.actor.data.spells.spell3.value, max: data.actor.data.spells.spell3.max },
-		4: { spells: [], value: data.actor.data.spells.spell4.value, max: data.actor.data.spells.spell4.max },
-		5: { spells: [], value: data.actor.data.spells.spell5.value, max: data.actor.data.spells.spell5.max },
-		6: { spells: [], value: data.actor.data.spells.spell6.value, max: data.actor.data.spells.spell6.max },
-		7: { spells: [], value: data.actor.data.spells.spell7.value, max: data.actor.data.spells.spell7.max },
-		8: { spells: [], value: data.actor.data.spells.spell8.value, max: data.actor.data.spells.spell8.max },
-		9: { spells: [], value: data.actor.data.spells.spell9.value, max: data.actor.data.spells.spell9.max }
-	}
-	
-	let spellCount = 0
-	let items = data.actor.items;
-
-	for (let item of items) {
-		if (item.type == "class") continue;
-		if (item.flags.favtab === undefined || item.flags.favtab.isFavourite === undefined) {
-			item.flags.favtab = { isFavourite: false };
-		}
-		let isFav = item.flags.favtab.isFavourite;
-		if (app.options.editable) {
-			let favBtn = $(`<a class="item-control item-fav" data-fav="${isFav}" title="${isFav ? "Remove from Favourites" : "Add to Favourites"}"><i class="${isFav ? "fas fa-bookmark" : "far fa-bookmark"}"></i></a>`);
-			favBtn.click(ev => {
-				app.actor.getOwnedItem(item._id).update({
-					"flags.favtab.isFavourite": !item.flags.favtab.isFavourite
-				});
-			});
-			html.find(`.item[data-item-id="${item._id}"]`).find('.item-controls').prepend(favBtn);
-		}
-		
-		if (isFav) {
-
-			item.quantity = item.data.quantity;
-			item.showquant = false;
-			if ( item.quantity != undefined && item.quantity > 1){
-				item.showquant = true;
-			}
-			
-			item.action = "";
-			if (item.data.activation) {
-				item.action = item.data.activation.cost +' '+item.data.activation.type;
-			} 
-			item.spellComps = "";
-			if (item.data.components) {
-				let comps = item.data.components;
-				let v = (comps.vocal) ? "V" : "";
-				let s = (comps.somatic) ? "S" : "";
-				let m = (comps.material) ? "M" : "";
-				let c = (comps.concentration) ? true : false;
-				let r = (comps.ritual) ? true : false;
-				item.spellComps = `${v}${s}${m}`;
-				item.spellCon = c;
-				item.spellRit = r;
-			}
-			
-			item.editable = app.options.editable;
-			switch (item.type) {
-			case 'feat':
-				if (item.flags.favtab.sort === undefined) {
-					item.flags.favtab.sort = (favFeats.count + 1) * 100000; // initial sort key if not present
-				}
-				favFeats.push(item);
-				break;
-			case 'spell':
-				if (item.data.preparation.mode) {
-					item.spellPrepMode = ` (${CONFIG.DND5E.spellPreparationModes[item.data.preparation.mode]})`
-				}
-				if (item.data.level) {
-					favSpells[item.data.level].spells.push(item);
-				} else {
-					favSpells[0].spells.push(item);
-				}
-				spellCount++;
-				break;
-			default:
-				if (item.flags.favtab.sort === undefined) {
-					item.flags.favtab.sort = (favItems.count + 1) * 100000; // initial sort key if not present
-				}
-				favItems.push(item);
-				break;
-			}
-		}
-	}
-	
-	// Alter core CSS to fit new button
-	if (app.options.editable) {
-		html.find('.spellbook .item-controls').css('flex', '0 0 70px');
-		html.find('.inventory .item-controls, .features .item-controls').css('flex', '0 0 70px');
-		html.find('.favourite .item-controls').css('flex', '0 0 22px');
-	}
-	
-	let tabContainer = html.find('.favorites-target');
-	data.favItems = favItems.length > 0 ? favItems.sort((a, b) => (a.flags.favtab.sort) - (b.flags.favtab.sort)) : false;
-	data.favFeats = favFeats.length > 0 ? favFeats.sort((a, b) => (a.flags.favtab.sort) - (b.flags.favtab.sort)) : false;
-	data.favSpells = spellCount > 0 ? favSpells : false;
-	data.editable = app.options.editable;
-	
-	await loadTemplates(['modules/tidy5e-sheet/templates/item.hbs']);
-	let favtabHtml = $(await renderTemplate('modules/tidy5e-sheet/templates/template.hbs', data));
-	favtabHtml.find('.item-name h4').click(event => app._onItemSummary(event));
-	
-	if (app.options.editable) {
-		favtabHtml.find('.item-image').click(ev => app._onItemRoll(ev));
-		let handler = ev => app._onDragItemStart(ev);
-		favtabHtml.find('.item').each((i, li) => {
-			if (li.classList.contains("inventory-header")) return;
-			li.setAttribute("draggable", true);
-			li.addEventListener("dragstart", handler, false);
-		});
-		//favtabHtml.find('.item-toggle').click(event => app._onToggleItem(event));
-		favtabHtml.find('.item-edit').click(ev => {
-			let itemId = $(ev.target).parents('.item')[0].dataset.itemId;
-			app.actor.getOwnedItem(itemId).sheet.render(true);
-		});
-		favtabHtml.find('.item-fav').click(ev => {
-			let itemId = $(ev.target).parents('.item')[0].dataset.itemId;
-			let val = !app.actor.getOwnedItem(itemId).data.flags.favtab.isFavourite;
-			app.actor.getOwnedItem(itemId).update({
-				"flags.favtab.isFavourite": val
-			});
-		});
-		
-		// Sorting
-		favtabHtml.find('.item').on('drop', ev => {
-			ev.preventDefault();
-			ev.stopPropagation();
-			
-			let dropData = JSON.parse(ev.originalEvent.dataTransfer.getData('text/plain'));
-			if (dropData.actorId !== app.actor.id || dropData.data.type === 'spell') return;
-			let list = null;
-			if (dropData.data.type === 'feat') list = favFeats;
-			else list = favItems;
-			let dragSource = list.find(i => i._id === dropData.data._id);
-			let siblings = list.filter(i => i._id !== dropData.data._id);
-			let targetId = ev.target.closest('.item').dataset.itemId;
-			let dragTarget = siblings.find(s => s._id === targetId);
-			
-			if (dragTarget === undefined) return;
-			const sortUpdates = SortingHelpers.performIntegerSort(dragSource, {
-				target: dragTarget,
-				siblings: siblings,
-				sortKey: 'flags.favtab.sort'
-			});
-			const updateData = sortUpdates.map(u => {
-				const update = u.update;
-				update._id = u.target._id;
-				return update;
-			});
-			app.actor.updateEmbeddedEntity("OwnedItem", updateData);
-		});
-	}
-	tabContainer.append(favtabHtml);
-	try {
-		if (game.modules.get("betterrolls5e") && game.modules.get("betterrolls5e").active) BetterRolls.addItemContent(app.object, favtabHtml, ".item .item-name h4", ".item-properties", ".item > .rollable div");
-	} 
-	catch (err) {
-		// Better Rolls not found!
-	}
-	Hooks.callAll("renderedTidy5eSheet", app, html, data);
-
-}
-
 // Preload tidy5e Handlebars Templates
 Hooks.once("init", () => {
   preloadTidy5eHandlebarsTemplates();
@@ -372,10 +232,10 @@ Actors.registerSheet("dnd5e", Tidy5eSheet, {
 });
 
 Hooks.on("renderTidy5eSheet", (app, html, data) => {
+	addFavorites(app, html, data, scrollPos);
 	migrateTraits(app, html, data);
 	addClassList(app, html, data);
 	setSheetClasses(app, html, data);
-	addFavorites(app, html, data);
 	checkDeathSaveStatus(app, html, data);
 	hidePortraitButtons(app, html, data);
 	// console.log(data);
